[Feat] #56 ReviewRegisterCell을 버튼으로 분리, keyboard 애니메이션 추가
